# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "CycleCheckResponse": {
        "additionalProperties": false,
        "properties": {
          "scheduler": {
            "$ref": "#/components/schemas/SchedulerStatus"
          },
          "triggered": {
            "title": "Triggered",
            "type": "boolean"
          }
        },
        "required": [
          "triggered",
          "scheduler"
        ],
        "title": "CycleCheckResponse",
        "type": "object"
      },
      "CycleTriggerResponse": {
        "additionalProperties": false,
        "properties": {
          "message": {
            "title": "Message",
            "type": "string"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          }
        },
        "required": [
          "ok",
          "message"
        ],
        "title": "CycleTriggerResponse",
        "type": "object"
      },
...
      "ErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "items": {
                  "$ref": "#/components/schemas/ValidationIssue"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "error": {
            "title": "Error",
            "type": "string"
          },
          "ok": {
            "const": false,
            "default": false,
            "title": "Ok",
            "type": "boolean"
          }
        },
        "required": [
          "error",
          "code"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
      "ExecutedOrderResponse": {
        "additionalProperties": false,
        "properties": {
          "action": {
            "enum": [
              "BUY",
              "SELL"
            ],
            "title": "Action",
            "type": "string"
          },
          "cash_after": {
            "title": "Cash After",
            "type": "number"
          },
          "fee": {
            "title": "Fee",
            "type": "number"
          },
          "price": {
            "title": "Price",
            "type": "number"
          },
          "quantity": {
            "title": "Quantity",
            "type": "number"
          },
          "ticker": {
            "title": "Ticker",
            "type": "string"
          },
          "total": {
            "title": "Total",
            "type": "number"
          }
        },
        "required": [
          "ticker",
          "action",
          "quantity",
          "price",
          "total",
          "fee",
          "cash_after"
        ],
        "title": "ExecutedOrderResponse",
        "type": "object"
      },
...
      "InstrumentSummary": {
        "additionalProperties": false,
        "properties": {
          "company": {
            "title": "Company",
            "type": "string"
          },
          "instrument_type": {
            "title": "Instrument Type",
            "type": "string"
          },
          "ticker": {
            "title": "Ticker",
            "type": "string"
          }
        },
        "required": [
          "ticker",
          "company",
          "instrument_type"
        ],
        "title": "InstrumentSummary",
        "type": "object"
      },
...
      "ManualTradePreviewRequest": {
        "additionalProperties": false,
        "properties": {
          "action": {
            "enum": [
              "BUY",
              "SELL"
            ],
            "title": "Action",
            "type": "string"
          },
          "amount_dollars": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "maximum": 1000000.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,10}|(?=[\\d.]{1,13}0*$)\\d{0,10}\\.\\d{0,2}0*$)",
                "type": "string"
              }
            ],
            "title": "Amount Dollars"
          },
          "ticker": {
            "maxLength": 10,
            "minLength": 1,
            "pattern": "^[A-Za-z.\\-]+$",
            "title": "Ticker",
            "type": "string"
          },
          "username": {
            "default": "taavet",
            "pattern": "^[A-Za-z][A-Za-z0-9_]{1,19}$",
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "ticker",
          "action",
          "amount_dollars"
        ],
        "title": "ManualTradePreviewRequest",
        "type": "object"
      },
      "ManualTradeRequest": {
        "additionalProperties": false,
        "properties": {
          "action": {
            "enum": [
              "BUY",
              "SELL"
            ],
            "title": "Action",
            "type": "string"
          },
          "amount_dollars": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "maximum": 1000000.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,10}|(?=[\\d.]{1,13}0*$)\\d{0,10}\\.\\d{0,2}0*$)",
                "type": "string"
              }
            ],
            "title": "Amount Dollars"
          },
          "client_order_id": {
            "format": "uuid",
            "title": "Client Order Id",
            "type": "string"
          },
          "ticker": {
            "maxLength": 10,
            "minLength": 1,
            "pattern": "^[A-Za-z.\\-]+$",
            "title": "Ticker",
            "type": "string"
          },
          "username": {
            "default": "taavet",
            "pattern": "^[A-Za-z][A-Za-z0-9_]{1,19}$",
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "ticker",
          "action",
          "amount_dollars",
          "client_order_id"
        ],
        "title": "ManualTradeRequest",
        "type": "object"
      },
...
      "OrderPreviewResponse": {
        "additionalProperties": false,
        "properties": {
          "action": {
            "enum": [
              "BUY",
              "SELL"
            ],
            "title": "Action",
            "type": "string"
          },
          "cash_before": {
            "title": "Cash Before",
            "type": "number"
          },
          "current_holding_quantity": {
            "title": "Current Holding Quantity",
            "type": "number"
          },
          "current_holding_value": {
            "title": "Current Holding Value",
            "type": "number"
          },
          "current_holding_weight": {
            "title": "Current Holding Weight",
            "type": "number"
          },
          "estimated_cash_after": {
            "title": "Estimated Cash After",
            "type": "number"
          },
          "estimated_executable_amount": {
            "title": "Estimated Executable Amount",
            "type": "number"
          },
          "estimated_holding_quantity": {
            "title": "Estimated Holding Quantity",
            "type": "number"
          },
          "estimated_holding_value": {
            "title": "Estimated Holding Value",
            "type": "number"
          },
          "estimated_holding_weight": {
            "title": "Estimated Holding Weight",
            "type": "number"
          },
          "estimated_quantity": {
            "title": "Estimated Quantity",
            "type": "number"
          },
          "fee": {
            "title": "Fee",
            "type": "number"
          },
          "instrument": {
            "$ref": "#/components/schemas/InstrumentSummary"
          },
          "max_buy_amount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Buy Amount"
          },
          "max_sell_amount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Sell Amount"
          },
          "quote": {
            "$ref": "#/components/schemas/QuoteResponse"
          },
          "requested_amount": {
            "title": "Requested Amount",
            "type": "number"
          },
          "unrealized_pnl": {
            "title": "Unrealized Pnl",
            "type": "number"
          },
          "warnings": {
            "items": {
              "$ref": "#/components/schemas/OrderWarningResponse"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "instrument",
          "quote",
          "action",
          "requested_amount",
          "estimated_executable_amount",
          "estimated_quantity",
          "fee",
          "cash_before",
          "estimated_cash_after",
          "current_holding_quantity",
          "current_holding_value",
          "estimated_holding_quantity",
          "estimated_holding_value",
          "current_holding_weight",
          "estimated_holding_weight",
          "max_buy_amount",
          "max_sell_amount",
          "unrealized_pnl",
          "warnings"
        ],
        "title": "OrderPreviewResponse",
        "type": "object"
      },
      "OrderWarningResponse": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "OrderWarningResponse",
        "type": "object"
      },
...
      "QuoteResponse": {
        "additionalProperties": false,
        "properties": {
          "change_percent": {
            "title": "Change Percent",
            "type": "number"
          },
          "price": {
            "title": "Price",
            "type": "number"
          },
          "timestamp": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timestamp"
          }
        },
        "required": [
          "price",
          "change_percent",
          "timestamp"
        ],
        "title": "QuoteResponse",
        "type": "object"
      },
...
      "ResetResponse": {
        "additionalProperties": false,
        "properties": {
          "message": {
            "title": "Message",
            "type": "string"
          },
          "ok": {
            "const": true,
            "title": "Ok",
            "type": "boolean"
          }
        },
        "required": [
          "ok",
          "message"
        ],
        "title": "ResetResponse",
        "type": "object"
      },
      "SchedulerResult": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "stocks_processed": {
            "minimum": 0.0,
            "title": "Stocks Processed",
            "type": "integer"
          }
        },
        "required": [
          "stocks_processed",
          "error"
        ],
        "title": "SchedulerResult",
        "type": "object"
      },
      "SchedulerStatus": {
        "additionalProperties": false,
        "properties": {
          "in_progress": {
            "title": "In Progress",
            "type": "boolean"
          },
          "last_result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SchedulerResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "last_run": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Run"
          },
          "next_run": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Run"
          },
          "running": {
            "title": "Running",
            "type": "boolean"
          }
        },
        "required": [
          "running",
          "last_run",
          "next_run",
          "in_progress",
          "last_result"
        ],
        "title": "SchedulerStatus",
        "type": "object"
      },
...
      "TradeResponse": {
        "additionalProperties": false,
        "properties": {
          "ok": {
            "const": true,
            "title": "Ok",
            "type": "boolean"
          },
          "transaction": {
            "$ref": "#/components/schemas/ExecutedOrderResponse"
          }
        },
        "required": [
          "ok",
          "transaction"
        ],
        "title": "TradeResponse",
        "type": "object"
      },
...
      "ValidationIssue": {
        "additionalProperties": false,
        "properties": {
          "location": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "location",
          "message",
          "type"
        ],
        "title": "ValidationIssue",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Portfolio Simulator",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/cycle": {
      "post": {
        "operationId": "trigger_cycle_api_cycle_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CycleTriggerResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "Trigger Cycle",
        "tags": [
          "operations"
        ]
      }
    },
    "/api/cycle/check": {
      "post": {
        "operationId": "check_cycle_api_cycle_check_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CycleCheckResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Operator access required"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "Check Cycle",
        "tags": [
          "operations"
        ]
      }
    },
    "/api/cycle/status": {
      "get": {
        "operationId": "cycle_status_api_cycle_status_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SchedulerStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "Cycle Status",
        "tags": [
          "operations"
        ]
      }
    },
...
    "/api/reset": {
      "post": {
        "description": "Wipe all portfolios \u2014 reset cash to $10K, clear holdings and transactions.",
        "operationId": "reset_portfolios_api_reset_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResetResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "Reset Portfolios",
        "tags": [
          "operations"
        ]
      }
    },
...
    "/api/trade": {
      "post": {
        "operationId": "execute_api_trade_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManualTradeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TradeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid operation"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Operator access required"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Operation conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "Execute",
        "tags": [
          "trades"
        ]
      }
    },
    "/api/trade/preview": {
      "post": {
        "operationId": "preview_api_trade_preview_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManualTradePreviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderPreviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid operation"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Operator access required"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Resource not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "Preview",
        "tags": [
          "trades"
        ]
      }
    },
...
  }
}